+++ tests/Traits/AdminMethodsDestroyTest.php
@@ -3,6 +3,7 @@
 namespace Larrock\Core\Tests\Traits;
 
 use DaveJamesMiller\Breadcrumbs\BreadcrumbsServiceProvider;
+use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Request;
 use Larrock\ComponentAdminSeo\LarrockComponentAdminSeoServiceProvider;
 use Larrock\ComponentBlocks\BlocksComponent;
@@ -69,6 +70,9 @@ class AdminMethodsDestroyTest extends TestCase
         $this->assertCount(1, $test->shareMethods());
     }
 
+    /**
+     * @throws \Exception
+     */
     public function testDestroy()
     {
         $request = new Request();
@@ -76,6 +80,63 @@ class AdminMethodsDestroyTest extends TestCase
         $load = $test->destroy($request, 1);
         $this->assertEquals(302, $load->getStatusCode());
         $this->assertNull(Blocks::find(1));
+        $this->assertEquals('http://localhost', $load->getTargetUrl());
+
+        //Удаление с параметром category_item
+        \DB::connection()->table('blocks')->insert([
+            'title' => 'test',
+            'description' => 'test',
+            'url' => 'test',
+        ]);
+        $request = Request::create('/admin/destroy', 'DELETE', [
+            'category_item' => 'category'
+        ]);
+        /** @var RedirectResponse $load */
+        $load = $test->destroy($request, 2);
+        $this->assertEquals('http://localhost/admin/blocks/category', $load->getTargetUrl());
+        $this->assertEquals(302, $load->getStatusCode());
+
+        //Удаление с параметром place=material
+        \DB::connection()->table('blocks')->insert([
+            'title' => 'test',
+            'description' => 'test',
+            'url' => 'test',
+        ]);
+        $request = Request::create('/admin/destroy', 'DELETE', [
+            'place' => 'material'
+        ]);
+        /** @var RedirectResponse $load */
+        $load = $test->destroy($request, 3);
+        $this->assertEquals('http://localhost/admin/blocks', $load->getTargetUrl());
+        $this->assertEquals(302, $load->getStatusCode());
+
+        //Удаление группы материалов
+        \DB::connection()->table('blocks')->insert([
+            'title' => 'test2',
+            'description' => 'test',
+            'url' => 'test2',
+        ]);
+        \DB::connection()->table('blocks')->insert([
+            'title' => 'test3',
+            'description' => 'test',
+            'url' => 'test3',
+        ]);
+        $request = Request::create('/admin/destroy', 'DELETE', [
+            'ids' => [4,5]
+        ]);
+        /** @var RedirectResponse $load */
+        $load = $test->destroy($request, 2);
+        $this->assertCount(0, Blocks::all());
+        $this->assertEquals(302, $load->getStatusCode());
+        $this->assertEquals('http://localhost', $load->getTargetUrl());
+
+        //Удаление несуществующего элемента
+        $request = new Request();
+        /** @var RedirectResponse $load */
+        $load = $test->destroy($request, 100);
+        $this->assertEquals('Такого материала уже не существует', $load->getSession()->get('message.danger.0'));
+        $this->assertEquals(302, $load->getStatusCode());
+        $this->assertEquals('http://localhost', $load->getTargetUrl());
     }
 }