+++ tests/Traits/AdminMethodsShowTest.php
@@ -3,19 +3,22 @@
 namespace Larrock\Core\Tests\Traits;
 
 use DaveJamesMiller\Breadcrumbs\BreadcrumbsServiceProvider;
-use Larrock\ComponentBlocks\BlocksComponent;
-use Larrock\ComponentBlocks\LarrockComponentBlocksServiceProvider;
+use Illuminate\Http\Request;
+use Illuminate\View\View;
+use Larrock\ComponentCategory\LarrockComponentCategoryServiceProvider;
+use Larrock\ComponentFeed\FeedComponent;
+use Larrock\ComponentFeed\LarrockComponentFeedServiceProvider;
 use Larrock\Core\LarrockCoreServiceProvider;
-use Larrock\Core\Tests\DatabaseTest\CreateBlocksDatabase;
+use Larrock\Core\Tests\DatabaseTest\CreateCategoryDatabase;
+use Larrock\Core\Tests\DatabaseTest\CreateFeedDatabase;
 use Larrock\Core\Tests\DatabaseTest\CreateMediaDatabase;
-use Larrock\Core\Tests\DatabaseTest\CreateSeoDatabase;
-use Larrock\Core\Traits\AdminMethodsEdit;
+use Larrock\Core\Traits\AdminMethodsShow;
 use Larrock\Core\Traits\ShareMethods;
 use Orchestra\Testbench\TestCase;
 use Proengsoft\JsValidation\JsValidationServiceProvider;
 use Spatie\MediaLibrary\MediaLibraryServiceProvider;
 
-class AdminMethodsEditTest extends TestCase
+class AdminMethodsShowTest extends TestCase
 {
     protected function getEnvironmentSetUp($app)
     {
@@ -31,11 +34,11 @@ class AdminMethodsEditTest extends TestCase
     {
         parent::setUp();
 
-        $seed = new CreateBlocksDatabase();
-        $seed->setUpBlocksDatabase();
+        $seed = new CreateFeedDatabase();
+        $seed->setUpFeedDatabase();
 
-        $seed = new CreateSeoDatabase();
-        $seed->setUpSeoDatabase();
+        $seed = new CreateCategoryDatabase();
+        $seed->setUpCategoryDatabase();
 
         $seed = new CreateMediaDatabase();
         $seed->setUpMediaDatabase();
@@ -45,7 +48,8 @@ class AdminMethodsEditTest extends TestCase
     {
         return [
             LarrockCoreServiceProvider::class,
-            LarrockComponentBlocksServiceProvider::class,
+            LarrockComponentCategoryServiceProvider::class,
+            LarrockComponentFeedServiceProvider::class,
             BreadcrumbsServiceProvider::class,
             JsValidationServiceProvider::class,
             MediaLibraryServiceProvider::class
@@ -55,36 +59,37 @@ class AdminMethodsEditTest extends TestCase
     protected function getPackageAliases($app)
     {
         return [
-            'LarrockBlocks' => 'Larrock\ComponentBlocks\Facades\LarrockBlocks',
+            'LarrockCategory' => 'Larrock\ComponentCategory\Facades\LarrockCategory',
+            'LarrockFeed' => 'Larrock\ComponentFeed\Facades\LarrockFeed',
             'Breadcrumbs' => 'DaveJamesMiller\Breadcrumbs\Facades\Breadcrumbs'
         ];
     }
 
     public function testShareMethods()
     {
-        $test = new AdminMethodsEditMock();
+        $test = new AdminMethodsShowMock();
         $this->assertCount(1, $test->shareMethods());
     }
 
-    public function testEdit()
+    public function testShow()
     {
-        $test = new AdminMethodsEditMock();
-        $test = $test->edit(1);
-        $this->assertEquals('larrock::admin.admin-builder.edit', $test->getName());
-        $this->assertEquals('test', $test->getData()['data']->title);
-        $this->assertNotEmpty($test->render());
+        $test = new AdminMethodsShowMock();
+        /** @var View $load */
+        $load = $test->show(1);
+        $this->assertEquals('larrock::admin.admin-builder.categories', $load->getName());
+        $this->assertNotEmpty($load->render());
     }
 }
 
-class AdminMethodsEditMock
+class AdminMethodsShowMock
 {
-    use AdminMethodsEdit, ShareMethods;
+    use AdminMethodsShow, ShareMethods;
 
     protected $config;
 
     public function __construct()
     {
-        $this->config = new BlocksComponent();
+        $this->config = new FeedComponent();
         $this->config->shareConfig();
     }
 }