+++ views/admin/admin-builder/categories.blade.php
@@ -1,18 +1,18 @@
 @extends('larrock::admin.main')
-@section('title') {{ $app->name }} admin @endsection
+@section('title') {{ $package->name }} admin @endsection
 
 @section('content')
     <div class="container-head uk-margin-bottom">
         <div class="uk-grid">
             <div class="uk-width-expand">
                 @if(isset($category))
-                    {!! Breadcrumbs::render('admin.'. $app->name .'.category', $category) !!}
+                    {!! Breadcrumbs::render('admin.'. $package->name .'.category', $category) !!}
                     <div class="uk-clearfix"></div>
                     <a class="link-blank" href="{{ $category->full_url }}">{{ $category->full_url }}</a>
                 @else
-                    {!! Breadcrumbs::render('admin.'. $app->name .'.index') !!}
+                    {!! Breadcrumbs::render('admin.'. $package->name .'.index') !!}
                     <div class="uk-clearfix"></div>
-                    <a class="link-blank" href="/{{ $app->name }}">/{{ $app->name }}</a>
+                    <a class="link-blank" href="/{{ $package->name }}">/{{ $package->name }}</a>
                 @endif
             </div>
             <div class="uk-width-auto">
@@ -23,7 +23,7 @@
                     @if(isset($allowCreate))
                         <a href="#add_category" class="uk-button uk-button-primary uk-width-1-1 uk-width-auto@s show-please" data-target="create-category"
                            data-focus="create-category-title">Добавить подраздел</a>
-                        <a class="uk-button uk-button-primary uk-width-1-1 uk-width-auto@s" href="/admin/{{ $app->name }}/create?category={{ $category->id }}">Добавить материал</a>
+                        <a class="uk-button uk-button-primary uk-width-1-1 uk-width-auto@s" href="/admin/{{ $package->name }}/create?category={{ $category->id }}">Добавить материал</a>
                     @endif
                 @else
                     @if(isset($allowCreate))
@@ -53,9 +53,9 @@
                 </thead>
                 <tbody class="uk-sortable" data-uk-sortable="{handleClass:'uk-sortable-handle'}">
                 @if(isset($category))
-                    @include('larrock::admin.category.include-create-easy', ['parent' => $category->id, 'component' => $app->name])
+                    @include('larrock::admin.category.include-create-easy', ['parent' => $category->id, 'component' => $package->name])
                 @else
-                    @include('larrock::admin.category.include-create-easy', ['parent' => 0, 'component' => $app->name])
+                    @include('larrock::admin.category.include-create-easy', ['parent' => 0, 'component' => $package->name])
                 @endif
                 @if(count($categories) === 0)
                     <div class="uk-alert uk-alert-warning">Разделов еще нет</div>
@@ -83,15 +83,15 @@
                         <th class="uk-text-right">
                             @if(isset($category))
                                 @php if($category->parent === 0){ $category->parent = NULL; } @endphp
-                                <a href="{{ $category->parent or '/admin/'. $app->name }}">
+                                <a href="{{ $category->parent or '/admin/'. $package->name }}">
                                     <span class="uk-margin-small-right" uk-icon="reply"></span>..
                                 </a>
                             @endif
                         </th>
-                        @if(isset($app->rows['title']))
-                            <th>{{ $app->rows['title']->title }}</th>
+                        @if(isset($package->rows['title']))
+                            <th>{{ $package->rows['title']->title }}</th>
                         @endif
-                        @foreach($app->rows as $row)
+                        @foreach($package->rows as $row)
                             @if($row->inTableAdmin || $row->inTableAdminEditable)
                                 <th style="width: 90px" class="uk-visible@s">{{ $row->title }}</th>
                             @endif
@@ -105,14 +105,14 @@
                     @else
                         <tr class="tr-massiveAction">
                             <td colspan="6">
-                                @include('larrock::admin.admin-builder.massive-action', ['data' => $data, 'app' => $app->name, 'formId' => 'massiveActionMaterials'])
+                                @include('larrock::admin.admin-builder.massive-action', ['data' => $data, 'app' => $package->name, 'formId' => 'massiveActionMaterials'])
                             </td>
                         </tr>
                         @foreach($data as $data_value)
                             <tr>
                                 <td width="55">
                                     <div class="actionSelect" data-target="massiveActionMaterials" data-id="{{ $data_value->id }}">
-                                        @if($app->plugins_backend && array_key_exists('images', $app->plugins_backend)
+                                        @if($package->plugins_backend && array_key_exists('images', $package->plugins_backend)
                                         && $image = $data_value->getMedia('images')->sortByDesc('order_column')->first())
                                             <img style="width: 55px" src="{{ $image->getUrl('110x110') }}">
                                         @else
@@ -120,27 +120,27 @@
                                         @endif
                                     </div>
                                 </td>
-                                @if(isset($app->rows['title']))
+                                @if(isset($package->rows['title']))
                                     <td>
-                                        <a class="uk-h4" href="/admin/{{ $app->name }}/{{ $data_value->id }}/edit">{{ $data_value->title }}</a>
+                                        <a class="uk-h4" href="/admin/{{ $package->name }}/{{ $data_value->id }}/edit">{{ $data_value->title }}</a>
                                         <br/>
                                         <a class="link-to-front" target="_blank" href="{{ $data_value->full_url }}" title="ссылка на элемент на сайте">
                                             {{ str_limit($data_value->full_url, 35, '...') }}
                                         </a>
                                     </td>
                                 @endif
-                                @foreach($app->rows as $row)
+                                @foreach($package->rows as $row)
                                     @if($row->inTableAdminEditable)
                                         @if($row instanceof \Larrock\Core\Helpers\FormBuilder\FormCheckbox)
                                             <td class="row-active uk-visible@s">
                                                 <div class="uk-button-group btn-group_switch_ajax" role="group" style="width: 100%">
                                                     <button type="button" class="uk-button uk-button-primary uk-button-small
                                                             @if($data_value->{$row->name} === 0) uk-button-outline @endif"
-                                                            data-row_where="id" data-value_where="{{ $data_value->id }}" data-table="{{ $app->table }}"
+                                                            data-row_where="id" data-value_where="{{ $data_value->id }}" data-table="{{ $package->table }}"
                                                             data-row="active" data-value="1" style="width: 50%">on</button>
                                                     <button type="button" class="uk-button uk-button-danger uk-button-small
                                                             @if($data_value->{$row->name} === 1) uk-button-outline @endif"
-                                                            data-row_where="id" data-value_where="{{ $data_value->id }}" data-table="{{ $app->table }}"
+                                                            data-row_where="id" data-value_where="{{ $data_value->id }}" data-table="{{ $package->table }}"
                                                             data-row="active" data-value="0" style="width: 50%">off</button>
                                                 </div>
                                             </td>
@@ -149,7 +149,7 @@
                                                 <input type="text" value="{{ $data_value->{$row->name} }}" name="{{ $row->name }}"
                                                        class="ajax_edit_row form-control uk-input uk-form-small" data-row_where="id"
                                                        data-value_where="{{ $data_value->id }}"
-                                                       data-table="{{ $app->table }}">
+                                                       data-table="{{ $package->table }}">
                                                 @if($row->name === 'position')
                                                     <i class="uk-sortable-handle uk-icon uk-icon-bars uk-margin-small-right" title="Перенести материал по весу"></i>
                                                 @endif
@@ -158,7 +158,7 @@
                                             <td class="uk-visible@s">
                                                 <select class="ajax_edit_row form-control uk-select uk-form-small" data-row_where="id"
                                                         data-value_where="{{ $data_value->id }}"
-                                                        data-table="{{ $app->table }}" data-row="{{ $row->name }}">
+                                                        data-table="{{ $package->table }}" data-row="{{ $row->name }}">
                                                     @foreach($row->options as $option)
                                                         <option @if($option === $data_value->{$row->name}) selected @endif value="{{ $option }}">{{ $option }}</option>
                                                     @endforeach
@@ -190,7 +190,7 @@
     @if(isset($category->getChild))
         @if(count($category->getChild) === 0)
             <table class="uk-table uk-table-striped">
-                @include('larrock::admin.category.include-create-easy', ['parent' => $category->id, 'component' => $app->name])
+                @include('larrock::admin.category.include-create-easy', ['parent' => $category->id, 'component' => $package->name])
             </table>
         @else
             <div class="uk-margin-large-bottom ibox-content">
@@ -200,7 +200,7 @@
                         <th>
                             @if(isset($category))
                                 @php if($category->parent === 0){ $category->parent = NULL; } @endphp
-                                <a href="{{ $category->parent or '/admin/'. $app->name }}">
+                                <a href="{{ $category->parent or '/admin/'. $package->name }}">
                                     <span class="uk-margin-small-right" uk-icon="reply"></span>..
                                 </a>
                             @endif
@@ -215,7 +215,7 @@
                     </tr>
                     </thead>
                     <tbody class="uk-sortable" data-uk-sortable="{handleClass:'uk-sortable-handle'}">
-                    @include('larrock::admin.category.include-create-easy', ['parent' => $category->id, 'component' => $app->name])
+                    @include('larrock::admin.category.include-create-easy', ['parent' => $category->id, 'component' => $package->name])
                     @include('larrock::admin.category.include-list-categories', ['data' => $category->getChild])
                     </tbody>
                 </table>